pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "error-fix-engine"
version = "1.0.0"
description = "Automated SEO engine that generates and publishes error-fix articles"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.5.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "markdown>=3.5.2",
    "python-slugify>=8.0.1",
    "tenacity>=8.2.3",
    "structlog>=24.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
